+++ test/mitmproxy/addons/test_browser.py
@@ -27,3 +27,33 @@ async def test_no_browser():
         with taddons.context() as tctx:
             b.start()
             await tctx.master.await_log("platform is not supported")
+
+
+async def test_get_browser_cmd_executable():
+    with mock.patch("shutil.which") as which:
+        which.side_effect = lambda cmd: cmd == "chrome"
+        assert browser.get_browser_cmd() == ["chrome"]
+
+
+async def test_get_browser_cmd_no_executable():
+    with mock.patch("shutil.which") as which:
+        which.return_value = False
+        assert browser.get_browser_cmd() is None
+
+
+async def test_get_browser_cmd_flatpak():
+    def subprocess_run_mock(cmd, **kwargs):
+        returncode = 0 if cmd == ["flatpak", "info", "com.google.Chrome"] else 1
+        return mock.Mock(returncode=returncode)
+
+    with mock.patch("shutil.which") as which, mock.patch("subprocess.run") as subprocess_run:
+        which.side_effect = lambda cmd: cmd == "flatpak"
+        subprocess_run.side_effect = subprocess_run_mock
+        assert browser.get_browser_cmd() == ["flatpak", "run", "-p", "com.google.Chrome"]
+
+
+async def test_get_browser_cmd_no_flatpak():
+    with mock.patch("shutil.which") as which, mock.patch("subprocess.run") as subprocess_run:
+        which.side_effect = lambda cmd: cmd == "flatpak"
+        subprocess_run.return_value = mock.Mock(returncode=1)
+        assert browser.get_browser_cmd() is None